Convert to tsx.
